Use Java types for TypeNames, Kotlin types for processing
This CL changes KspProcessingEnv to always return the Kotlin
representation of a type even when it is a java specific type
(e.g. getTypeElement is called with java.lang.String, we'll
return the KspType for kotlin.String).
This makes it consistent with the java compiler where the
compiler will always return the kotlin type even when the
qualified name is explicitly specified as the java type or if
the type is originating from a java source file.

As Room generates java code, we'll be generating TypeNames in
Java. This will (in theory :) ) allow Room code generation work
as usual while allowing the rest of the processing to take
advantage of Kotlin types.

Right now, KSP does not do the java to kotlin mapping for us.
(see https://github.com/google/ksp/issues/126). Until that issue
is resolved, I've added a temporary KspTypeMapper class to do
the conversion.

This change should allow merging more of the XProcessing common
tests but I've not touched them to minimize the CL which is already
big.

There are also two known cases which are not handled in this CL.
One major case is when the return type of a function is void.
Right now, they are all mapped to kotlin.Unit because kotlin.Unit
exists in java type system as well and sometimes it leaks into
the java code generated by KAPT. We need to narrow down when it
happens and hopefully workaround.

Second case is when a generic value is overridden. I'm not sure
if that will be a problem because in that case, kotlin seems to
generate both methods and Room already handles it (aosp/1441551).

There is a good chance that we can resolve both of these issues by
detecting overrides but haven't tried it.

After this change, we should be able to start running integration
tests to catch the real issues ¯\_(ツ)_/¯. That being said, I think
the model of using Kotlin during processing and Java for final type
generation is probably the right way to proceed forward.

Bug: 160322705
Test: existing tests
Change-Id: I3885bc51807d29a9be13955ae6f12bafcdab3dc0