Fix crash when ComposeView is zero sized
Drawing of AndroidComposeView which was measured with 0x0 size was causing a crash as we were not considering zero sized constraints as a valid ones and the LayoutNode were not measured before the drawing was requested for it. I also fixed the fact that AbstractComposeView was unnecessary clipping the inner AndroidComposeView so now all the compose views could draw outside of its bounds.

Test: added a test to AndroidLayoutDrawTest to verify zero sized ComposeView can correctly draw outside of its bounds and not crash
Fixes: 169230467
Change-Id: I50d3408a38ff5dfe7fe71ccb85b227cd31269928