Ongoing Activities library, Java version.
Ongoing activities refer to user initiated & time bound activities that
could be running in the background of the watch (workouts, calls, media,
etc). This is a new API for handling ongoing activities that provides a
rich experience for the developer, using Notifications as a
communication mechanism but leaving the door open to change that in the
future.

Relnote: New API for Ongoing Activities, this is a no-op on "unsupported devices."
Test: Manual - automated tests on a follow up cl

Change-Id: I69a31a74749eeb6d290d7943892019a5a1cdfd62