Add presubmit support to androidx suite config
This change makes it so that different options/tests
get executed in presubmit vs postsubmit

It makes a small change to the AMD to better support
the use case. This is because I needed to be able to
distinguish between when we build all because of no
changes (postsubmit) vs when we build all because of
unknown files (e.g. presubmit, change to buildSrc)

Test: Added unit test support for xml verification, but mostly manual
instpection of output

Change-Id: I39815c3efa4d1c72f19fdd7923c54d2d5b0fe2e7