Replace Context with explicit Activity references.
This chagnes the ExtensionInterface to accept Activity references as
opposed to Context referneces in line with API feedback to use Activity
references until other such Contexts are supported. This also updates
the use of Context references in other parts of the library.

Relnote: "Updates ExtensionInterface to accept explicity Activity
references."

Test: All WM tests
Fixes: 167350560
Change-Id: I07ded6e81dd3d1006adc11c8b66dccd9caa3000d