Improve kotlin array support
This CL fixes a couple of issues in KspArray implementation.

Previously, it completely ignored all built in array types
which meant classes like kotlin.IntArray were not considered
as arrays. Fixed it.
Also, we would generate int[] for Array<Int> which is wrong
as kotlin always boxes them (e.g. Array<java.lang.Integer>).
Fixed it as well.

Now KspArray is a sealed class with two implementations for
primitives and non-primitives.

Added tests for both cases that also run with KAPT to ensure
consistency.

Bug: 160322705
Test: XArrayTypeTest
Change-Id: I4efefe3f44f09885709fca46f041cec7ecd96678