TextField RTL Layout Fix
Defining the min height and min width in the TextField
layout ends up in component having a different width than
the text.

In this CL, the constraints that are passed to
TextFieldDelegate.layout is determined beforehand so that the
text can cover a wider area  and text can be aligned correctly.
This brings back the old behavior with the min width constraint.

Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:connectedAndroidTest
Test: ./gradlew compose:material:material:test
Test: ./gradlew compose:material:material:connectedAndroidTest

RelNote: N/A

Bug: 172510134
Change-Id: I81ea79ac321b6ddc8fd04785a2fb7c6820cb6fd8