Remove usages of ConflatedBroadcastChannel
This CL removes all usages of ConflatedBroadcastChannel.

For some cases, all we wanted was a state flow hence I replaced it with
StateFlow. For other cases, we use it like an event bus where StateFlow
does not work as is since it does not allow dispatching the same value
twice OR does not allow not having a default value.
For those cases, I've added a ConflatedEventBus class which allows not
having an initial value and also does allow sending the same value twice.

Bug: 171708323
Test: existing tests, ConflatedEventBusTest
Change-Id: Idba2dc099f2d734df486eb14157c5f3791106b9d