WatchFace: Add support for Direct Boot
We need to seralize the parameters WCS used to initalize the watch
face and keep the style updated. In a direct boot scenario were the
user is locked and WCS won't connect until much later we should
load these params and init the watch face.

Also includes changes to de flake WatchFaceControlClientTest.

Bug: 173092459
Test: New golden image test
Change-Id: If93bff4b9b01027a32ca170656b842a969e223f4