Allow REFRESH requests to act like retry
Previously, when a request failed, RMA always expected a retry call
before ever running that request type.

This becomes an issue w/ pull to refresh where user likely means retry
the calls (because we explicity request it only in that case or in first
try).

This CL changes REFRESH requests to ignore load errors.
It is not completely equal to retryFailed because we don't clear
possible error states for append prepend UNLESS the new refresh
succeeds.

Bug: 173438474
Test: RemoteMediatorAccessorTest
Change-Id: I7d6c684c4ce2f336c6a8c253759fbdae905ee1ce