Remove page size from LoadParams
This CL removes pageSize from LoadParams. PageSize is confusing and was
already deprecated.

It is now set by the PageFetcher / PagedList when we detect the paging
source is legacy. This creates an issue in tests where pager is not
involved. When that happens, we print a warning and guess it from the
load parameters.

Bug: 149157296
Test: existing tests
Relnote: "LoadParams.pageSize has been removed (it was already
deprecated).
The recommendation is to use `LoadParams.loadSize` in your PagingSource.

`LoadParams.loadSize` is always equal to the `PagingConfig.pageSize`
except for the initial load call where it is equal to the
`PagingConfig.initialLoadSize`.

If you are testing your Paging2 DataSource without using a Pager or
PagedList, `pageSize` may not match the `PagingConfig.pageSize` if
you are also setting `initialLoadSize`. If it is important for your
tests, try using a Pager/PagedList instead which will internally set
the correct PageSize for your DataSource load methods."

Change-Id: I98ac735090f65919282c4afbd005bbd27b9af0f3