Ignore certain directories when generating release notes
In almost all cases, changes to these directory types
are for internal purposes and not something we need to
document in our release notes.

The most common case is that a change to library A
will require changes in samples or tests that use library A
but sit along side library B.  However, this doesn't mean
that library B has actually changed so we should generate
release notes as such.

Test: ./generateReleaseNotes.py 1605117600000
Change-Id: I627443bf6bfeca87f95e1d09af4db004830884bd