Port IndexBasedArrayIteratorTest.kt back to Java
IndexBasedArrayIteratorTest.kt is a JUnit test written in Kotlin. When we
eventually port all tests to Kotlin, I think it's beneficial to run them as
native Kotlin tests.

In addition, backporting IndexBasedArrayIteratorTest to Java allows us to
revive IndexBasedArrayIteratorTest.kt soon and run the Kotlin tests in parallel
to the Java tests.

Test: The backported Java test class

Change-Id: I7a38f6e0bc8bc86b5a83abc401b626d7c53d5fbf