Stub out compat classes for Zoom and EvComp
This change refactors several core classes in the CameraPipe integration
library to add basic stubs for adapting Zoom and EvComp to parameters in
used by Camera2.

The main functional change is that ZoomState and ExposureCompensation
interfaces are now only implemented by a kotlin data class, which makes
the state easier to reason about and test.

Test: Built and ran the CameraX Core Test App

Change-Id: I7b578f2b3ee6fb485bbf7095a92a7e3ddeef3684