Remove direct dispatcher
This CL removes the direct dispatcher and either uses a built in one if
it does not matter for the test or uses the available test dispatchers.

I'm not 100% sure on a few of them where we control the execution but
seems fine ?

The current implementation of Direct dispatcher was not fitting the
documentation on CoroutineDispatcher. More specifically, dispatch method
should not immediately run the given runnable:
https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-coroutine-dispatcher/dispatch.html

Bug: 174622674
Test: existing test
Change-Id: Ida63553b110bc6a4ff740b791903c29b6775a186