Add CompositionData to replace SlotTable for tooling API usage
Introduced CompositionData and CompositionGroup as the public
API the tooling API uses to read internal composition state.
This allows removing SlotTable, SlotReader and SlotWriter from
the public API.

Test: CompositionDataTests, ui-tooling tests
Relnote: """Removed removed SlotTable, SlotReader and
SlotWriter from the public API. These were marked as
InternalComposeAPI previously. Now they are internal
to the compose module.

CompositionData and CompositionGroup where added as a
replacement for the ui-tooling API to use to extract
composition information. These are public but are not
intended for use outside the ui-tooling API as they provide
the raw information the ui-tooling API interprets"""

Change-Id: I31a9ca6a7e5bbf162c984394dffd7a25b059315a