Fix broken unlock3a tests
There is no blocking intermediate step in unlock3A method.
To unlock 3a we need to send the needed captured requests right
away without checking the 3a state or waiting for a camera response.

Thus if we do the below sequence :
1. val res = async { unlock3A }
2. assertThat(res.isCompleted).isFalse()

(2) can fail or pass depending on how quickly the unlock3A
method go executed.

So it's unnecessary to check for the completion of the async
task, but instead we should check the Deferred<Result3A> returned
by unlock3A method(which we are already doing).

Test: multiple iterations of
      ./gradlew :camera:camera-camera2-pipe:test

Change-Id: Ia68e9fb304f169183c545cb0b1a570aa6d03c99e