Ensure ActivityResultRegistry keeps saved pending results
When an ActivityResultRegistry has its state saved, we should ensure
the state in the returned bundle is unchanged. Since the pending results
were being saved directly, unregistering a result after the saving the
state would cause pending results to be dropped.

We should instead save a copy of the pending result in the bundle to
ensure that the state that is restored is the same state that was saved.

RelNote: "The ActivityResultRegistry now always restores the exact state
that was saved."
Test: testUnregisterAfterSavedState

Change-Id: Idd56b95f825aa457d623a9569ce229d42c7f746b