Remove identityHashCode() public api
We only used it two cases:
When we override hashCode() in some classes in animations module: there it so point to it as it is already a default implementation of Object.hashCode().
When we wanted to have hashCode to have some stable ordering when we sort LayoutNodes. We again can just do layoutNode.hashCode() relying on a default implementation in Object.hashCode().

Fixes: 175124473
Test: existing tests
Relnote: Removed Any.identityHashCode() public api
Change-Id: I025d720aec64ebd2182787b9200ca9b3827d5436