fix multi step type converters
This CL fixes a bug in TypeAdapterStore where we mixed up assignability
order when checking whether a type converter satisfies an output type,
causing it to either miss valid cases or returning false adapters.
For instance, it would miss an assignment from ArrayList to List but
accept assignment from List to ArrayList.

I've made two other changes:
* When looking for matches from candidates, we'll prioritize exact match
over assignable match. This is still not perfect because we don't visit
all candidates in queue before returning assignable match. In a
followup, we can change this to do a weighted BFS that visits all
possible nodes that are in the same level and add costs to assignable
matches.
* Instead of using isAssignableWithoutVariance, we now use isAssignable.
I'm not 100% sure on that but seems like the right method call though
I'm not sure if it will hit some weird KAPT case that is not covered in
tests. Still, it is better to use isAssignable until we find a failing
case (the test in the original CL which added withoutVariance passes
with this fix)

Bug: 175916061
Test: TypeAdapterStoreTest#multiStepTypeConverters

Change-Id: I7a750ada653ffa82dce477e2af3bbf5bf46e026d