Handle primitives overriding generics in KSP
When a kotlin class specifies a non-nullable primitive type for a
super's generic argument, kotlin will duplicate methods that receive the
generic as an argument.
This CL changes the KSP override check to ignore those overrides so that
it also reports two methods. Note that this only happens for paremeters
and not return values. For return values, only the boxed one is generated.

To handle the return type boxing, we now use the overridden desclaration
when wrapping types, which lets us choose between primitive and
non-primitive by looking at the declaration.

Bug: 160258066
Bug: 160322705
Test: XExecutableElementTest

Change-Id: Id2fb76748ff014bcabe4543daed211d82d04bd4c