Move XType.is checks to extension methods
This CL moves XType.is checks for known types to extension methods and
uses the TypeName (thats what JavacType did).

Initially, they were put into the interface because I wasn't sure how
well the typeName will work in KSP. At this point, it is one of the more
reliable parts of the KSP implementation (e.g. we do very well in
deciding whether we want it to be a primitive or not).

By moving them to extension methods, KSP and Javac implementations get
closer and becomes more consistent.

While creating a test case (KSP impl used to fail to detect boxed Void),
I hit a bug in KSP. Added a workaround for now.
https://github.com/google/ksp/issues/200

Bug: 160322705
Test: XTypeTest & existing tests
Change-Id: I2ece2a9738faeef9a8530475f9f96330b35abaf3