[KSP] Handle single values for array fields in annotations
In java code, if an array value is set without using the array syntax
(e.g. not wrapped in {}), KSP might return it as a single item instead
of a list of items.
https://github.com/google/ksp/issues/214

This CL adds a workaround to assume it when a single item is received
when we are expecting an array and wraps it into a list.

Bug: 176180385
Test: XAnnotationBoxTest.typeReferenceArray_singleItemInJava
Test: TypeAdapterStoreTest.testInvalidNonStaticInnerClass

Change-Id: Ia5d2513f28550eace91831cf9213a046dafae020