ImageAnalysis: acquire image when queue is maxed should not crash.
Not sure what the exact cause of the bug b/175851631, but this CL is a step in the right direction. If the image queue is maxed out, acquiring a new image should not crash. This is a net improvment on user experience. It doesn't hide the issue either because if it happens preview is going to freeze.

Changes:
- refactor abstract analyzerto make the logic clear. Remove the open/close methods of the abstract analyzer, and replace it with the later introduced attach/detach concept. It should only forward images when 1) the user analzyer is set, and 2) the ImageAnalysis is attached to a camera.
- Add a clearCache() method to release cached images. The method is called when 1) user analyzer is unset, 2) ImageAnalysis is detached, or 3) there is an error in creating pipeline.
- In non-blocking analyzer, rely on posted image's existence to check if it's been closed.
- try/catch the max queue exception and log it as an error.
- clean up unit tests.

Relnote: This change catches the exception when the image queue is maxed out in ImageAnalysis. So instead of crash, you may notice other use cases being frozen or sluggish. e.g. frozen/sluggish preview.

Bug: 175851631
Test: manual test and ./gradlew bOS
Change-Id: Ic12dafa0b0feed875a07ea1d8c23c52977b15afa