Fix flaky gesturescope.sendClickTest
The sendClickTest creates some test boxes and then clicks two of them. It has a ClickRecorder modifier that records the clicks.

My change aosp/1541805 introduced a feature where the view gets focus when the user clicks on any whitespace in the app.
With this change, on SDK 27 the first click in sendClickTest was recorded, this click caused the view to gain focus. After that the second click was not recorded.
However, on SDK 28, both the clicks were recorded.

This CL reverts the feature that allows a user to click on white space in the app to request view focus. It uses an alternate fix for b/170188271 so we don't need this feature for now.
However in the future if we need this feature we will have to revisit this test.

Bug: 176953083
Bug: 170188271
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.test.gesturescope.SendClickTest
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.FocusManagerTest
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.OwnerFocusTest
Change-Id: I544b76bd9c7985121b369d311c317528d6f980f1