Ignore special effects for views not attachedToWindow
When adding fragments to the activity before the View hierarchy has been
attached to the window, if the fragment has special effects, they will
never be started. In the case of Transitions, this means there is never
an onEndTransition() callback, which causes fragment to be unable to
move to resumed.

Since unattached views will not run their transitions anyway, we should
drop any special effect operations before the special effects container
is attached to the window.

RelNote: "Fragments with transitions added before the view hierarchy is
attached now properly reach resumed."
Test: FragmentTransitionTest
Bug: 177154873

Change-Id: I1fc1db01450466f9ee012f07e4b608aa3337c838