Provide APIs for retrieving @HiltViewModel from Navigation in Compose scopes.
Relnote: Create a new artifact, ':hilt-navigation-compose' with APIs, `hiltNavGraphViewModel()` that allows users to get a @HiltViewModel annotated ViewModel from a Navigation back stack entry within a Compose application using ':navigation-compose'. The APIs can either take a `NavBackStackEntry` or a 'route' string that can be used to scope the ViewModel to either the current back stack entry or to a parent entry in the stack.
Test: ./gradlew :hilt:hilt-navigation-compose:cAT
Change-Id: Ia92343c281823099390b2e1795b89cdb1abfc440