Add more complex statuses.
Ongoing Activities now support more complex statuses, they are composed
of a template and 'parts' to use to fill the placeholders on the
template.
Placeholders in the template have the form '#name#', and '##' can be
used in a template to produce a single '#' in the output.
Statuses can have more than one template, to be used in the future when
we introduce new 'part' classes that are not supported in older
versions.

Test: Added
Relnote: "Added support for more complex status. They are composed of a
template (or several) and a series of Parts that will be used to fill the
template's placeholders"

Change-Id: I1fe818ee09171f618699cce0b5046bec0bc79e91