Add semantics role to image and icon
In View system there is no special classname for Icon as there is no separate widget. I decided to use Role.Image for Icon as well because when Icon is not actionable it is essentially an image. Alternatively considered creating a role description for Icon though I don't see much benefit.

Bug: 149735981
Test: new unit tests, checked demo with talkback on
Relnote: N/A
Change-Id: I7f4ebe3b546ab5e2c6e832feafe522f03a039d7a