Add API for saving the back stack
The Fragment back stack is made up of separate
BackStackRecords, also known as transactions (as they
implement FragmentTransaction and that's the public
API used to create them). While a popBackStack()
operation will reverse the transaction and destroy
all of the no longer referenced fragments, to *save*
the state we need a new type of operation.

This operation is called saveBackStack() and takes the
name of the back stack entry (the one set via
addToBackStack()). This required converting some
common code (finding the correct index on the back stack)
so that it is useable by both pop and save.

This does not yet actually do any saving, nor is the
restore half of the equation built yet, but this is
the first step down that path.

Test: new SaveRestoreBackStackTest suite
BUG: 80029773
Relnote: N/A
Change-Id: I661f33c0ea7bb67338f7d9e75f8e8b8454d2bc21