AutoMigration initial CL, handling the simplest schema change, adding a new column.
In the current implementation, the AutoMigration code will be able to detect the addition of a new column to a schema. It will also be able to detect if a column has been deleted (or renamed, can't difference between these two in this implementation) or if the affinity of a column has been changed. In these cases, the processor will output an error message that includes the name of the column in reference. In the follow-up CLs to this CL, we will aiming to handle all simple schema changes.

Test: SchemaDifferTest.kt has a list of schema bundles that test whether or not an automigration can be generated with the given schemas of a database. These tests all contain the same valid "from" schema (version 1), and different "to" schemas that test valid input as well as invalid input that output an expected error message. AutoMigrationProcessorTest.kt will have more cases in follow up CLs.

Test: AutoMigrationProcessorTest.kt, SchemaDifferTest.kt
Bug: 74441348
Change-Id: Ib164fee76e25cacb6e57746d0128bf0ffba5eefa