Disable the failing datastore stress tests.
From https://android-build.googleplex.com/builds/treetop/android-review/1562487?ref=COMMENT&revision=3&workplanId=L78300000800639067 it is clear that it is these two tests that are causing the time outs. I am disabling them here for now until I can figure out a good way to test these (maybe disable disk IO for these tests?) so people aren't blocked on this.

It's not 100% clear to me whether these are just taking longer on the server so they are timing out or if they are actually broken somehow.

Test: N/A
Bug: 178641154
Relnote: N/A
Change-Id: If23a6f2a48f32470765af5a2af1c66ac861335d3