Add getter to match builder setter.
Currently, validTimeRange is a private property of the base class, as
a helper method (isActiveAt) is provided for the developer's
convenience. However, this is asymmetric as the builder for some of
the subtypes can be used to set this field but, once constructed, the
field is no longer accessible.

This commit moves the field to the derived classes which provide a way
to set this field.

Bug: 171788713
Test: ./gradlew :wear:wear-watchface-complications-rendering:assembleDebug :wear:wear-complications-provider:assembleDebug :wear:wear-watchface-data:assembleDebug :wear:wear-watchface:assembleDebug :wear:wear-watchface-client:assembleDebug
Relnote: Expose the validTimeRange of ComplicationData.
Change-Id: I91366d280466c1139b02cae4ed53b32a53ca5446