Rename CATEGORY_PROVIDER_CONFIG_ACTION.
This commit renames CATEGORY_PROVIDER_CONFIG_ACTION to become
CATEGORY_PROVIDER_CONFIG because the constant does not represent an action but
a category. While it is associated with METADATA_KEY_PROVIDER_CONFIG_ACTION, it
is misleading to have action in its name.

Bug: 171788617
Test: ./gradlew :wear:wear-watchface-complications-rendering:assembleDebug :wear:wear-complications-provider:assembleDebug :wear:wear-watchface-data:assembleDebug :wear:wear-watchface:assembleDebug :wear:wear-watchface-client:assembleDebug
Relnote: Rename CATEGORY_PROVIDER_CONFIG_ACTION to CATEGORY_PROVIDER_CONFIG.
Change-Id: I7c0689956ef6533ed3ac16edcc0487038cc96a55