Changes Indication#createInstance to be @Composable
Changes LocalIndication to just contain the type Indication instead of a lambda returning Indication

Indication can now just be treated as a factory for IndicationInstance and there do not need to be multiple Indication instances across the hierarchy. createInstance() can query theme CompositionLocals directly for any theme information that depends on where the Modifier.indication() is invoked in the hierarchy.

Also changes Surface to provide ambients outside the LayoutNode with modifiers - this means that ripples applied directly to the Surface will now consume the content color provided by the Surface, instead of the nearest parent provider outside the Surface.

Bug: b/157150564
Test: MaterialRippleThemeTest
Test: SurfaceTest
Relnote: "Changed Indication#createInstance to be @Composable, and changed LocalIndication to contain an Indication, not () -> Indication."
Change-Id: I5eeea2424e4deda6116f0b48b690b628f8d545eb