Ensure saveBackStack() is self contained
A key part to saving a back stack is to be
able to later restore that same back stack.

As such, we need to ensure that the
FragmentTransactions that are being saved
are self contained - i.e., that the only
fragments they specifically mention are
added as part of those transactions. By
ensuring that the saved transactions are
self contained, we can ensure that they
can be correctly restored even if the
underlying state of the FragmentManager
is changed between the save and restore
operations.

This builds upon the previous topmost
fragment change so as to continue to allow
operations like replace() which don't
explicitly mention previously added fragments.

As part of this, generateOpsForPendingActions() has
been made more resilient to crashes - if a
generateOps() call crashes, it still throws the
exception, but the set of pending actions that
crashed FragmentManager will now be cleared out
so that they are not reprocessed any time later
if something (like our test suite...) catches
exceptions thrown by executePendingOperations().

Test: new savePreviouslyReferencedFragment test
BUG: 80029773
Change-Id: I0bde3c27f415990c2918d98a79ea76c3d0c6b55e