UserStyle wire formats to not use inner classes
It seems libcore in Android has a different implementation than
openJDK and class getSimpleName can return different results. This
can break versioned parcelable for inner classes so we refactor to
not use them in wire formats.

Test: Manual testing
Bug: 179127569
Relnote: We've fixed a marshalling bug with UserStyle wireformats changing some of the unstable hidden API.
Change-Id: I8be0900af78624868d3d7fbaeac0e7752d5beda1